Move freeze into system startup and rollback into After hook
This is tricky because if the box were started outside this system in
an unfrozen state, the code as written would not notice that. need to
think on that

also rollback and possibly freeze doesn't seem to be actually
working, just noticed, but suspect the bug's been there for some time.